requirements.txt
python-telegram-bot>=21.0
groq>=0.4.2
neo4j>=5.17.0
astrapy>=1.2.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.4
python-docx>=1.1.0
Pillow>=10.2.0
pyvis>=0.3.2
networkx>=3.2.1
matplotlib>=3.8.3
python-dotenv>=1.0.1
httpx>=0.27.0
tenacity>=8.2.3
loguru>=0.7.2
numpy>=1.26.4
pydantic>=2.6.3
pydantic-settings>=2.2.1
aiofiles>=23.2.1
aiohttp>=3.9.3
